requirements.txt
langchain==0.1.0
langchain-community==0.0.10
manim==0.18.0
python-dotenv==1.0.0
pydantic==2.5.0
ollama==0.1.6
openai==1.6.1
anthropic==0.8.1
groq==0.4.1